Create wravindranpassage2
The early months of 1969 were a difficult period for Harrison: he had quit the band temporarily, he was arrested for marijuana possession, and he had his tonsils removed.









Harrison states in his autobiography, I, Me, Mine:







"Here Comes the Sun" was written at the time when Apple was getting like school, where we had to go and be businessmen: 'Sign this' and 'sign that.' Anyway, it seems as if winter in England goes on forever, by the time spring comes you really deserve it. So one day I decided I was going to sag off Apple and I went over to Eric Clapton's house. The relief of not having to go see all those dopey accountants was wonderful, and I walked around the garden with one of Eric's acoustic guitars and wrote "Here Comes the Sun".[2]







As Clapton states in his autobiography, the house in question, in Ewhurst, Surrey, is known as Hurtwood Edge.[3] When interviewed in the Martin Scorsese documentary George Harrison: Living in the Material World, Clapton said he believed the month was April. Data from two meteorological stations in the London area show that April 1969 set a record for sunlight hours for the 1960s. The Greenwich station recorded 189 hours for April, a high that was not beaten until 1984. The Greenwich data also show that February and March were much colder than the norm for the 1960s, which would account for Harrison's reference to a "long, cold, lonely winter".[4]