Final version of Mozaiques_bc.xml
I have added to the existing work the 5-9 facsimiles, as well as facsimile 2, and a comment about facsimile 1. I filled in the TEI header of Mozaiques_bc.xml  with data taken from the TEI header spreadsheet. If successfully uploaded, Katy will be able to add to this file her own transcribed and encoded facsimiles.